Expose Taxonomy and TaxonomyMetadata types
