+++ packages/js/product-editor/changelog/update-expose-tax-type
@@ -1,0 +1,4 @@
+Significance: minor
+Type: update
+
+Expose Taxonomy and TaxonomyMetadata types
