+++ packages/js/product-editor/src/blocks/generic/taxonomy/create-taxonomy-modal.tsx
@@ -21,7 +21,7 @@ import classNames from 'classnames';
 /**
  * Internal dependencies
  */
-import { Taxonomy } from './types';
+import { Taxonomy } from '../../../types';
 import useTaxonomySearch from './use-taxonomy-search';
 
 type CreateTaxonomyModalProps = {
