+++ packages/js/product-editor/src/blocks/generic/taxonomy/edit.tsx
@@ -19,9 +19,12 @@ import { useSelect } from '@wordpress/data';
  * Internal dependencies
  */
 import { CreateTaxonomyModal } from './create-taxonomy-modal';
-import { Taxonomy, TaxonomyMetadata } from './types';
 import useTaxonomySearch from './use-taxonomy-search';
-import { ProductEditorBlockEditProps } from '../../../types';
+import {
+	ProductEditorBlockEditProps,
+	Taxonomy,
+	TaxonomyMetadata,
+} from '../../../types';
 import useProductEntityProp from '../../../hooks/use-product-entity-prop';
 
 interface TaxonomyBlockAttributes extends BlockAttributes {
