--- packages/js/product-editor/src/blocks/generic/taxonomy/types.ts
@@ -1,9 +1,0 @@
-export interface Taxonomy {
-	id: number;
-	name: string;
-	parent: number;
-}
-
-export interface TaxonomyMetadata {
-	hierarchical: boolean;
-}
