+++ packages/js/product-editor/src/blocks/generic/taxonomy/use-taxonomy-search.ts
@@ -7,7 +7,7 @@ import { escapeHTML } from '@woocommerce/components';
 /**
  * Internal dependencies
  */
-import { Taxonomy } from './types';
+import { Taxonomy } from '../../../types';
 
 async function getTaxonomiesMissingParents(
 	taxonomies: Taxonomy[],
